[DR-2933] AzureIntegrationTest teardown would exit prematurely
When an AzureIntegrationTest run had 1+ snapshots to delete on teardown, the teardown method would not complete since we tried to remove items from a snapshot ID list while iterating over it via forEach.
Concretely, what this meant is that we would never get to the billing profile delete step in the teardown, and subsequent test runs in the same test suite would fail fast with a 'cannot reuse existing application deployment integrationapp from profile <x> with a different profile <y>' error.

- Updated the test teardown method to no longer remove elements from the snapshot ID list or null out instance-level resource IDs: these instance variables will all be reset in the test setup method.
- Moved contents of helper method clearEnvironment to test teardown method, and removed clearEnvironment: we should be handling this teardown in one place, rather than calling clearEnvironment at the end of test runs while also trying to clean up resources in the teardown method (which runs on both expected and premature test exits).
- Log billing profiles, snapshot summaries, and dataset summaries created in DataRepoFixtures for easier debugging in case of test error: none of these objects are large, and we might end up in a situation where we're debugging a test failure on a test environment that has since been repurposed for another test run.

+++ src/test/java/bio/terra/integration/AzureIntegrationTest.java
@@ -145,7 +145,6 @@ public class AzureIntegrationTest extends UsersBase {
     stewardToken = authService.getDirectAccessAuthToken(steward.getEmail());
     dataRepoFixtures.resetConfig(steward);
     profileId = dataRepoFixtures.createAzureBillingProfile(steward).getId();
-    datasetId = null;
     retryOptions =
         new RequestRetryOptions(
             RetryPolicyType.EXPONENTIAL,
@@ -165,24 +164,24 @@ public class AzureIntegrationTest extends UsersBase {
 
   @After
   public void teardown() throws Exception {
-    dataRepoFixtures.resetConfig(steward);
-    snapshotIds.forEach(
-        snapshotId -> {
-          try {
-            dataRepoFixtures.deleteSnapshot(steward, snapshotId);
-            snapshotIds.remove(snapshotId);
-          } catch (Exception ex) {
-            logger.warn("[Cleanup exception] Unable to delete snapshot " + snapshotId, ex);
-          }
-        });
-    if (datasetId != null) {
-      dataRepoFixtures.deleteDatasetLog(steward, datasetId);
-      datasetId = null;
-    }
+    logger.info(
+        "Teardown: trying to delete snapshots {}, dataset {}, billing profile {}",
+        snapshotIds,
+        datasetId,
+        profileId);
 
+    dataRepoFixtures.resetConfig(steward);
+
+    if (snapshotIds != null) {
+      for (UUID snapshotId : snapshotIds) {
+        dataRepoFixtures.deleteSnapshot(steward, snapshotId);
+      }
+    }
+    if (datasetId != null) {
+      dataRepoFixtures.deleteDataset(steward, datasetId);
+    }
     if (profileId != null) {
-      dataRepoFixtures.deleteProfileLog(steward, profileId);
-      profileId = null;
+      dataRepoFixtures.deleteProfile(steward, profileId);
     }
   }
 
@@ -311,9 +310,6 @@ public class AzureIntegrationTest extends UsersBase {
         // TODO: fix bug where this shows up as a 403 and not a 404 since it's not longer in Sam
         equalTo(403));
     assertThrows(AssertionError.class, () -> dataRepoFixtures.deleteProfile(steward, profileId));
-
-    // Make sure that any failure in tearing down is presented as a test failure
-    clearEnvironment();
   }
 
   @Test
@@ -922,7 +918,6 @@ public class AzureIntegrationTest extends UsersBase {
 
     // Make sure that any failure in tearing down is presented as a test failure
     blobIOTestUtility.deleteContainers();
-    clearEnvironment();
   }
 
   @Test
@@ -1039,7 +1034,6 @@ public class AzureIntegrationTest extends UsersBase {
 
     // Make sure that any failure in tearing down is presented as a test failure
     blobIOTestUtility.deleteContainers();
-    clearEnvironment();
   }
 
   @Test
@@ -1167,7 +1161,6 @@ public class AzureIntegrationTest extends UsersBase {
 
     // Make sure that any failure in tearing down is presented as a test failure
     blobIOTestUtility.deleteContainers();
-    clearEnvironment();
   }
 
   @Test
@@ -1231,8 +1224,6 @@ public class AzureIntegrationTest extends UsersBase {
         "allowing a bad row means 2 rows still succeeded",
         ingestResponseSuccess.getRowCount(),
         equalTo(2L));
-
-    clearEnvironment();
   }
 
   @Test
@@ -1284,28 +1275,6 @@ public class AzureIntegrationTest extends UsersBase {
         dataRepoFixtures.ingestJsonData(steward, datasetId, ingestRequest);
 
     dataRepoFixtures.assertCombinedIngestCorrect(ingestResponse, steward);
-
-    clearEnvironment();
-  }
-
-  // Make sure that any failure in tearing down is presented as a test failure
-  private void clearEnvironment() throws Exception {
-    for (UUID snapshotId : snapshotIds) {
-      if (snapshotId != null) {
-        dataRepoFixtures.deleteSnapshot(steward, snapshotId);
-        snapshotIds.remove(snapshotId);
-      }
-    }
-
-    if (datasetId != null) {
-      dataRepoFixtures.deleteDataset(steward, datasetId);
-      datasetId = null;
-    }
-
-    if (profileId != null) {
-      dataRepoFixtures.deleteProfile(steward, profileId);
-      profileId = null;
-    }
   }
 
   private String getSourceStorageAccountPrimarySharedKey() {
